Task 6done m

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
-monthly_income = int(input("Enter your monthly income: "))
-monthly_expenses = int(input("Enter your total monthly expenses: "))
-monthly_saving = monthly_income - monthly_expenses
+monthly_income = input("Enter your monthly income: ")
+monthly_expenses = input("Enter your total monthly expenses: ")
+monthly_saving = float(monthly_income) - float(monthly_expenses)
 projected_savings = int(monthly_saving * 12 + monthly_saving * 12 * 0.05)
 print(f"Your monthly savings are ${monthly_saving}.")
 print(f"Projected savings after one year, with interest, is: ${projected_savings}.")
